Task #6 - done
The user enters the month number from the keyboard (from
1 to 12). Depending on the month number received
the program displays the inscription: Winter (if entered
value 1.2 or 12), Spring (if a value from 3 is entered
to 5), Summer (if a value from 6 to 8 is entered), Autumn
(if a value from 9 to 11 is entered).
If the user entered a value not in the range of 1
up to 12 it is required to display an error message.